Move experiment orchestration to run execution
