+++ isaaclab_arena/tests/test_run_execution.py
@@ -10,7 +10,7 @@ import pytest
 
 from isaaclab_arena.environments.arena_environment_factory import ArenaEnvironmentCfg
 from isaaclab_arena.evaluation import run_execution
-from isaaclab_arena.evaluation.arena_run import ArenaRunCfg, RolloutLimitCfg, RunStatus
+from isaaclab_arena.evaluation.arena_run import ArenaRunCfg, ArenaRunResult, RolloutLimitCfg, RunStatus
 from isaaclab_arena.policy.policy_base import PolicyCfg
 
 
@@ -140,3 +140,63 @@ def test_build_and_run_requires_a_limit_for_an_unbounded_policy(monkeypatch, tmp
         )
 
     assert closed_resources == [(policy, environment)]
+
+
+def test_execute_experiment_runs_in_declaration_order(monkeypatch, tmp_path):
+    received = []
+
+    def build_and_run(run_cfg, output_dir, video_cfg):
+        received.append((run_cfg.name, output_dir, video_cfg.video_base_dir))
+        return ArenaRunResult(run_name=run_cfg.name, status=RunStatus.COMPLETED)
+
+    monkeypatch.setattr(run_execution, "build_and_run", build_and_run)
+
+    results = run_execution.execute_experiment([_run(name="first"), _run(name="second")], output_dir=tmp_path)
+
+    assert [result.run_name for result in results] == ["first", "second"]
+    assert received == [
+        ("first", str(tmp_path / "first"), str(tmp_path / "first")),
+        ("second", str(tmp_path / "second"), str(tmp_path / "second")),
+    ]
+
+
+def test_execute_experiment_records_failure_and_continues(monkeypatch, tmp_path):
+    attempted = []
+
+    def build_and_run(run_cfg, output_dir, video_cfg):
+        attempted.append(run_cfg.name)
+        if run_cfg.name == "failing":
+            raise RuntimeError("rollout failed")
+        return ArenaRunResult(run_name=run_cfg.name, status=RunStatus.COMPLETED)
+
+    monkeypatch.setattr(run_execution, "build_and_run", build_and_run)
+
+    results = run_execution.execute_experiment(
+        [_run(name="failing"), _run(name="passing")],
+        output_dir=tmp_path,
+        continue_on_error=True,
+    )
+
+    assert attempted == ["failing", "passing"]
+    assert [(result.run_name, result.status) for result in results] == [
+        ("failing", RunStatus.FAILED),
+        ("passing", RunStatus.COMPLETED),
+    ]
+
+
+def test_execute_experiment_stops_on_failure_by_default(monkeypatch, tmp_path):
+    attempted = []
+
+    def build_and_run(run_cfg, output_dir, video_cfg):
+        attempted.append(run_cfg.name)
+        raise RuntimeError("rollout failed")
+
+    monkeypatch.setattr(run_execution, "build_and_run", build_and_run)
+
+    with pytest.raises(RuntimeError, match="rollout failed"):
+        run_execution.execute_experiment(
+            [_run(name="failing"), _run(name="not_attempted")],
+            output_dir=tmp_path,
+        )
+
+    assert attempted == ["failing"]
